Make dendritejs work again
